refactorizado
