+++ .gitignore
@@ -1,0 +1,2 @@
+*.mp4
+__pycache__
